requirements.txt
# requirements.txt
Flask
pytesseract
opencv-python-headless
Pillow
numpy